"Improve Error Handling and URL Handling
Previously, when a user entered an incorrect post ID in the URL, the server returned an unclear error message. To enhance the user experience, I implemented ID validation, so now when a user types an invalid ID, they receive a clear error message indicating either an invalid post ID or a post not found (if the ID is of type ObjectId, but there's no matching post).

Additionally, I fixed the URL handling so that when a user enters 'posts/' in the URL, they will be directed to the posts page instead of displaying a post without any content.

These changes ensure a smoother user experience and better error feedback."